// tools/schematics/country-flags/schema.d.ts
// Copyright (c) 2019-2020 The Authors
// This file is the result of Happy Coding :-)
// https://github.com/abdes/happy-coding
//
// SPDX-License-Identifier: BSD-3-Clause

export interface Schema {
  countries: string[];
  name: string;
  project?: string;
  path?: string;
  all?: boolean;
  list?: boolean;
}